recall script hot-fix-label-scoring-export

See merge request my-archery/api!1608

+++ app/BLoC/Web/EventElimination/CleanEliminationMatch.php
@@ -120,17 +120,18 @@ class CleanEliminationMatch extends Retrieval
 
         $list_match = ArcheryEventEliminationGroupMatch::where("elimination_group_id", $elimination->id)->get();
         foreach ($list_match as $value) {
-            $group_team = ArcheryEventEliminationGroupTeams::find($value->group_team_id);
-            if ($group_team) {
-                $member_team = ArcheryEventEliminationGroupMemberTeam::where("participant_id", $group_team->participant_id)->get();
+            if ($value->group_team_id != 0) {
+                $group_team = ArcheryEventEliminationGroupTeams::find($value->group_team_id);
+                if ($group_team) {
+                    $member_team = ArcheryEventEliminationGroupMemberTeam::where("participant_id", $group_team->participant_id)->get();
 
-                foreach ($member_team as $mt) {
-                    $mt->delete();
+                    foreach ($member_team as $mt) {
+                        $mt->delete();
+                    }
+
+                    $group_team->delete();
                 }
-
-                $group_team->delete();
             }
-            
             $scoring = ArcheryScoringEliminationGroup::where("elimination_match_group_id", $value->id)->first();
             if ($scoring) {
                 $scoring->delete();
